backend/backend/requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.2
gunicorn==23.0.0
psycopg[binary]==3.2.7
celery==5.5.2
redis==6.0.0
djangorestframework==3.16.0
markdown==3.8
django-filter==25.1
#pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.1.0
coverage==7.9.2
psycopg2-binary==2.9.10